% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refresh.r
\name{asam_refresh}
\alias{asam_refresh}
\title{Refresh ASAM database}
\usage{
asam_refresh()
}
\description{
\code{asam_refresh} will fetch the latest (the files are udated daily)
             ASAM shapefile database and overwrite the one installed with the package
             (or previously updated).
}

