% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{asam_sf}
\alias{asam_sf}
\title{ASAM (Anti-shipping Activity Messages) Data (simple features)}
\format{A simple feature collection with 8,003 features and 12 fields}
\usage{
data(asam_sf)
}
\description{
This dataset contains data from the National Geospatial
             Intelligence Acency's Anti-shipping Activity Messages (ASAM)
             database. This include the locations and descriptive accounts
             of specific hostile acts against ships and mariners. The reports
             may be useful for recognition, prevention and avoidance of
             potential hostile activity.

\itemize{
  \item \code{Reference}: ASAM Reference Number
  \item \code{DateOfOcc}: Date of incident occurrence (a \code{Date} object)
  \item \code{SubReg}: Geographical Subregion. See
        \url{http://fas.org/irp/world/para/docs/geol_geo.htm} for a map of
        the boundaries for these locations or the \code{\link{asam_subregions}}
        function in this package.
  \item \code{Aggressor}: Who was the aggressor
  \item \code{Descript}: Description of the anti-shipping activity
  \item \code{Desc1}: Description of the anti-shipping activity
  \item \code{Desc2}: Description of the anti-shipping activity
  \item \code{Desc3}: Description of the anti-shipping activity
  \item \code{Desc4}: Description of the anti-shipping activity
  \item \code{Desc5}: Description of the anti-shipping activity
  \item \code{Desc6}: Description of the anti-shipping activity
}
}
\note{
Last updated 2018-09-23
}
\references{
\href{http://msi.nga.mil/NGAPortal/MSI.portal?_nfpb=true&_pageLabel=msi_portal_page_65}{ASAM home}
}
\seealso{
\code{\link{asam_refresh}} for updating the dataset,
}
\keyword{datasets}
