% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_asam}
\alias{read_asam}
\title{Read ASAM database

The ASAM database is updated daily by the National Geospatial-Intelligence
Agency (NGIA) and this function provides a simple means to download it.}
\usage{
read_asam(quiet = FALSE)
}
\arguments{
\item{quiet}{if `TRUE` no messages}
}
\description{
Read ASAM database

The ASAM database is updated daily by the National Geospatial-Intelligence
Agency (NGIA) and this function provides a simple means to download it.
}
\examples{
\dontrun{
read_asam()
}
}
