% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asam-package.r
\docType{package}
\name{asam}
\alias{asam}
\alias{asam-package}
\title{Access, download, update and work with Anti-shipping Activity Messages (ASAM)
database files}
\description{
Anti-shipping Activity Messages ('ASAM') include the locations and
descriptive accounts of specific hostile acts against ships and mariners.
The reports may be useful for recognition, prevention and avoidance of
potential hostile activity. A snapshot of the official 'ASAM' database is
included along with utilities to update the local database and the
official polygons defining the 'ASAM' global regions and subregions.
}
\details{
\itemize{
\item URL: \url{http://gitlab.com/hrbrmstr/asam}
\item BugReports: \url{https://gitlab.com/hrbrmstr/asam/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
