% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asam-package.r
\docType{package}
\name{asam}
\alias{asam}
\alias{asam-package}
\title{Access, download, update and work with Anti-shipping Activity Messages (ASAM)
database files}
\description{
Access, download, update and work with Anti-shipping Activity Messages (ASAM)
database files
}
\author{
Bob Rudis (@hrbrmstr)
}

