% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geo.r
\name{asam_subregions}
\alias{asam_subregions}
\title{Retrieve ASAM subregions polygons}
\usage{
asam_subregions()
}
\value{
\code{SpatialPolygonsDataFrame}
}
\description{
Returns the official ASAM Subregions polygons as a \code{SpatialPolygonsDataFrame}.
}
\details{
There are 9 ASAM global regions and 52 ASAM global subregions. These
are referenced in \code{REGION} and \code{SUBREGION} fields in the
returned object.
}
\examples{
subregions <- asam_subregions()
plot(subregions)
}

